requirements.txt
alembic==1.0.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.1
decorator==4.3.0
Fabric3==1.14.post1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
paramiko==2.4.2
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
PyNaCl==1.3.0
pytest==3.10.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
SQLAlchemy-Utils==0.33.6
validators==0.12.3
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-Alchemy==0.16.7
WTForms-Components==0.10.3
